+++ app/src/main/java/protect/card_locker/MainActivity.java
@@ -460,14 +460,6 @@ public class MainActivity extends CatimaAppCompatActivity implements LoyaltyCard
         groupsTabLayout.setVisibility(View.VISIBLE);
     }
 
-    private void openPrivacyPolicy() {
-        Intent browserIntent = new Intent(
-                Intent.ACTION_VIEW,
-                Uri.parse("https://catima.app/privacy-policy")
-        );
-        startActivity(browserIntent);
-    }
-
     @Override
     public boolean onCreateOptionsMenu(Menu inputMenu)
     {
@@ -577,12 +569,6 @@ public class MainActivity extends CatimaAppCompatActivity implements LoyaltyCard
         {
             Intent i = new Intent(getApplicationContext(), SettingsActivity.class);
             startActivityForResult(i, Utils.MAIN_REQUEST);
-            return true;
-        }
-
-        if(id == R.id.action_privacy_policy)
-        {
-            openPrivacyPolicy();
             return true;
         }