+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased - 85
 
+Android 4.4 is no longer supported starting with this release. If you want to use Catima on Android 4.4, please use version 2.6.1.
+
 - Improved Android 12 support
 - Improved about screen