+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Android 4.4 is no longer supported starting with this release. If you want to us
 
 - Improved Android 12 support
 - Improved about screen
+- Search now ignores accents
 
 ## v2.6.1 - 84 (2021-09-25)